CRUD Ciclos OK, CRUD RRHH en proceso

+++ src/main/java/com/example/demo/repository/CicloRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Ciclo;
 
 @Repository
-public interface CicloRepository extends JpaRepository<Ciclo, Long>{
+public interface CicloRepository extends JpaRepository<Ciclo, Integer>{
 	//public Ciclo findById(int id);
 	//public Ciclo findByNombre(String cicloName);
 }
